Move admin_flags from /utils to /admin
